+++ test/multi_json/adapter_selector/default_adapter_excluding_test.rb
@@ -68,7 +68,7 @@ class DefaultAdapterExcludingTest < Minitest::Test
   end
 
   def test_default_adapter_excluding_holds_mutex_during_detection
-    mutex = MultiJson::Concurrency.const_get(:DEFAULT_ADAPTER)
+    mutex = MultiJson::Concurrency.const_get(:MUTEXES).fetch(:default_adapter)
     synchronized = false
     mutex.define_singleton_method(:synchronize) do |&block|
       synchronized = true