+++ test/multi_json/core_methods/use/method_test.rb
@@ -138,7 +138,7 @@ class UseMethodTest < Minitest::Test # rubocop:disable Metrics/ClassLength
   end
 
   def test_use_holds_adapter_mutex_during_swap
-    mutex = MultiJson::Concurrency.const_get(:ADAPTER)
+    mutex = MultiJson::Concurrency.const_get(:MUTEXES).fetch(:adapter)
     calls = stub_synchronize_tracker(mutex)
     object = Class.new { include MultiJson }.new
     object.define_singleton_method(:load_adapter) { |arg| MultiJson.send(:load_adapter, arg) }